requirements.txt
absl-py==1.2.0
AMP==1.1.4
asgiref==3.4.1
cachetools==4.2.4
caldav==0.1.4
certifi==2022.6.15
charset-normalizer==2.0.12
coverage==6.2
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
Django==3.2.14
ffmpeg-python==0.2.0
future==0.18.2
google-auth==2.9.1
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h5py==2.10.0
idna==3.3
imageio==2.15.0
imgaug==0.4.0
importlib-metadata==4.8.3
kiwisolver==1.3.1
labelImg==1.8.6
lxml==4.9.1
Markdown==3.3.7
matplotlib==3.1.2
networkx==2.5.1
nose==1.3.7
numpy==1.17.0
oauthlib==3.2.0
onnx==1.2.3
onnxruntime-gpu==1.1.2
opencv-contrib-python==3.4.2.16
opencv-python==3.4.2.16
pandas==1.1.5
Pillow==8.4.0
prettytable==2.5.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
pytz==2022.1
PyWavelets==1.1.1
PyYAML==6.0
pyzmq==23.2.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.17.2
scipy==1.2.1
seaborn==0.11.2
Shapely==1.8.2
six==1.16.0
sqlparse==0.4.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2020.9.3
torch==1.7.1+cu110
torchaudio==0.7.2
torchsummary==1.5.1
torchvision==0.8.2+cu110
tqdm==4.60.0
typing==3.7.4.3
typing_extensions==4.1.1
urllib3==1.26.10
vobject==0.9.6.1
wcwidth==0.2.5
Werkzeug==2.0.3
wincertstore==0.2
zipp==3.6.0
zmq==0.0.0
